requirements.txt
Django==5.0.2
python-dotenv==1.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.61.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1
django-compressor==4.4
django-libsass==0.9
libsass==0.22.0
django-extensions==3.2.3
